Split out the proto and explain why with a comment
