+++ profiling/src/profile/pprof/mod.rs
@@ -1,2 +1,7 @@
+// Unless explicitly stated otherwise all files in this repository are licensed under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/). Copyright 2023-Present Datadog, Inc.
+
 mod proto;
+pub mod sliced_proto;
+
 pub use proto::*;
