+++ profiling/src/serializer/zipped_streaming_encoder.rs
@@ -1,6 +1,8 @@
-use std::io::Write;
+// Unless explicitly stated otherwise all files in this repository are licensed under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/). Copyright 2023-Present Datadog, Inc.
 
 use lz4_flex::frame::FrameEncoder;
+use std::io::Write;
 
 pub struct ZippedProtobufSerializer {
     buffer: Vec<u8>,
